feat(PatientProcess): split into components and hooks
